requirements.txt
fastapi
uvicorn
pyspark
xgboost
pandas
pyarrow
joblib
mlflow
kafka-python
